+++ cse/resources/cnt.js
@@ -5,6 +5,7 @@ const { generate_ri, get_cur_time, get_default_et, convert_loc_to_geoJson, get_l
 const sequelize = require('../../db/sequelize');
 
 const enums = require('../../config/enums');
+const { classify_create_error } = require('../create-error');
 const cin = require('./cin');
 const CIN = require('../../models/cin-model');
 const CNT = require('../../models/cnt-model');
@@ -99,8 +100,10 @@ async function create_a_cnt(req_prim, resp_prim) {
         resp_prim.pc = tmp_resp.pc;
     } catch (err) {
         logger.error({ err }, 'create_a_cnt failed');
-        resp_prim.rsc = enums.rsc_str['BAD_REQUEST'];
-        resp_prim.pc = { 'm2m:dbg': err.message };
+        // A name lost to a concurrent create is a conflict, not a bad request.
+        const { rsc, dbg } = classify_create_error(err);
+        resp_prim.rsc = rsc;
+        resp_prim.pc = { 'm2m:dbg': dbg };
     } finally {
         req_prim._pendingCreate?.resolve();
     }